# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ADM GO",
        "description": " API de aplicación ADM GO.",
        "contact": {
            "email": "basesinteligentes@birobid.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://181.198.213.18:8000",
            "description": "Server Pro"
        }
    ],
    "paths": {
        "/api/bancos": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de Bancos",
                "operationId": "App\\Http\\Controllers\\Banco::GetBancos",
                "responses": {
                    "200": {
                        "description": "Retorna codigo,nombre y estado de los bancos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/bacoscia": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de BancosCia",
                "operationId": "App\\Http\\Controllers\\BancoCiaController::GetBancoCia",
                "responses": {
                    "200": {
                        "description": "Retorna todos los registros de la tabla ADMBANCOCIA con ESTADO = 'A'."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/canton": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de Cantones",
                "operationId": "App\\Http\\Controllers\\CantonController::GetCantones",
                "responses": {
                    "200": {
                        "description": "Retorna todos los registros de la tabla ADMCANTON con ESTADO = 'A'."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/categoria": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de Categoria de Items",
                "operationId": "App\\Http\\Controllers\\CategoriaController::GetCategorias",
                "responses": {
                    "200": {
                        "description": "Retorna todos los registros de la tabla ADMCATEGORIA con ESTADO = 'A'."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/cliprecio": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de Precios",
                "operationId": "App\\Http\\Controllers\\CliPrecioController::GetCliPrecio",
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMCLIPRECIO."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/cliente": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Listado de Clientes  - Paginado 100.",
                "operationId": "App\\Http\\Controllers\\Cliente::listado",
                "responses": {
                    "200": {
                        "description": "Retorna todos los registros de la tabla ADMCLIENTE con ESTADO = 'A'."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/condicomer": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de Condiciones Comerciales",
                "operationId": "App\\Http\\Controllers\\CondiComercialesController::GetCondiComer",
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMCONDICOMER."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/condicomerxi/{codigo}": {
            "get": {
                "tags": [
                    "Datos"
                ],
                "summary": "Listado de Condiciones Comerciales de un Item",
                "operationId": "App\\Http\\Controllers\\CondiComercialesController::GetCondiComerPorProducto",
                "parameters": [
                    {
                        "name": "codigo",
                        "in": "path",
                        "description": "Codigo del Item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMCONDICOMER segun el codigo del item dado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/credito/{numero}": {
            "get": {
                "tags": [
                    "Credito"
                ],
                "summary": "Información de un credito segun el NUMCRE, donde TIPOCR = 'PAG' y TIPOCRE != 'ANT'",
                "operationId": "App\\Http\\Controllers\\CreditoController::GetCredito",
                "parameters": [
                    {
                        "name": "NUMCRE",
                        "in": "path",
                        "description": "numero",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMCREDITO segun el numero."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/creditopos/{numero}": {
            "get": {
                "tags": [
                    "Credito"
                ],
                "summary": "Información de un creditoPos segun el NUMCRE, donde TIPOCR = 'PAG' y TIPOCRE != 'ANT'",
                "operationId": "App\\Http\\Controllers\\CreditoController::GetCreditoPos",
                "parameters": [
                    {
                        "name": "NUMCRE",
                        "in": "path",
                        "description": "numero",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMCREDITOPOS segun el numero."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deuda": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDA donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB')",
                "operationId": "App\\Http\\Controllers\\DeudaController::GetDeudas",
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudaxv/{vendedor}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDA por vendedor, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB')",
                "operationId": "App\\Http\\Controllers\\DeudaController::GetDeudasXVendedor",
                "parameters": [
                    {
                        "name": "vendedor",
                        "in": "path",
                        "description": "vendedor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA por vendedor."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deuda/{cliente}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDA por cliente, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB')",
                "operationId": "App\\Http\\Controllers\\DeudaController::GetDeudaXCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA por cliente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/deudapos": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDAPOS, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB','CHP')",
                "operationId": "App\\Http\\Controllers\\DeudaPosController::GetDeudas",
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDA."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/deudapos/{cliente}": {
            "get": {
                "tags": [
                    "Deudas"
                ],
                "summary": "Retorna registros de ADMDEUDAPOS por Cliente, donde el saldo > 0, el ESTADO != null, y el TIPO in ('FAC','NVT','NDB','CHP')",
                "operationId": "App\\Http\\Controllers\\DeudaPosController::GetDeudaXCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "cliente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los registros de ADMDEUDAPOS y los Cheques de ADMDEUDA por cliente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}